requirements.txt
langgraph
langchain
fastapi
uvicorn
chromadb
google-generativeai
langchain-deepseek
langchain_google_genai
python-dotenv
python-socketio
pandas
openpyxl
tqdm